update: Cache를 Global/Infrastructure/Cache로 이동 및 Redis 구현 방식 변경
