--- PushAndPull/PushAndPull/Global/Cache/CacheStore.cs
@@ -1,47 +1,0 @@
-using StackExchange.Redis;
-using System.Text.Json;
-
-namespace PushAndPull.Global.Cache;
-
-public class CacheStore : ICacheStore
-{
-    private readonly IDatabase _db;
-
-    public CacheStore(IConnectionMultiplexer redis)
-    {
-        _db = redis.GetDatabase();
-    }
-
-    public async Task SetAsync<T>(string key, T value, TimeSpan? ttl = null)
-    {
-        var json = JsonSerializer.Serialize(value);
-
-        await _db.StringSetAsync(
-            (RedisKey)key,
-            (RedisValue)json,
-            ttl,
-            When.Always,
-            CommandFlags.None
-        );
-    }
-
-    public async Task<T?> GetAsync<T>(string key)
-    {
-        var value = await _db.StringGetAsync(key);
-
-        if (value.IsNullOrEmpty)
-            return default;
-
-        return JsonSerializer.Deserialize<T>(value!);
-    }
-
-    public async Task DeleteAsync(string key)
-    {
-        await _db.KeyDeleteAsync(key);
-    }
-
-    public async Task<bool> ExistsAsync(string key)
-    {
-        return await _db.KeyExistsAsync(key);
-    }
-}
