+++ PushAndPull/PushAndPull/Global/Config/GlobalServiceConfig.cs
@@ -1,4 +1,4 @@
-using PushAndPull.Global.Cache;
+using PushAndPull.Global.Infrastructure.Cache;
 using PushAndPull.Global.Service;
 
 namespace PushAndPull.Global.Config;
@@ -7,7 +7,7 @@ public static class GlobalServiceConfig
 {
     public static IServiceCollection AddGlobalServices(this IServiceCollection services)
     {
-        services.AddScoped<ICacheStore, CacheStore>();
+        services.AddSingleton<ICacheStore, CacheStore>();
         services.AddSingleton<IRoomCodeGenerator, RoomCodeGenerator>();
         services.AddSingleton<IPasswordHasher, PasswordHasher>();
         return services;
