+++ PushAndPull/PushAndPull/PushAndPull.csproj
@@ -25,7 +25,7 @@
         </PackageReference>
         <PackageReference Include="Npgsql" Version="10.0.1" />
         <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="9.0.4" />
-        <PackageReference Include="StackExchange.Redis" Version="2.10.1" />
+        <PackageReference Include="Microsoft.Extensions.Caching.StackExchangeRedis" Version="9.0.3" />
     </ItemGroup>
 
     <ItemGroup>
